requirements.txt
inquirer
databricks-sdk
typer
rich
sh